Create basic structure for mazestruct module
We are only going to initialize one maze object, but it's important enough to put in its own
module because the maze structure is how the avatars share information. The information contained
within each mazesquare is vital for the avatars to decide which way to walk each time it is
their turn. By keeping track of the walls on each of the sides of each square, we can slowly
start to figure out the structure of the maze, and the shared data structure allows us to use
that information for all of the avatars.

There is a getter and a setter method for each of the wall types because the avatars will be
looking up the information they and other avatars have gathered about each square as well as
entering more information into the shared pool whenever they "learn" it.