Add Makefile and .h for Mazestruct
Since Mazestruct is a helper module, it needs to have a .h file
to export its functions. Because it is modular, it's better to test it
separately from the other parts of the lab, so it needs its own makefile
to run a test file.Add Makefile and .h for Mazestruct

Since Mazestruct is a helper module, it needs to have a .h file
to export its functions. Because it is modular, it's better to test it
separately from the other parts of the lab, so it needs its own makefile
to run a test file.