Prepare for testing and fix compile errors/segfaults
Now, in the mazestruct module, you can compile and test the module using
simply `make` and `make test`, the latter of which will run an extremely
rudimentary bash script that calls a c testing script, which currently does nothing
except test creation, deletion, and printing of a maze.

There are segfaults in all but maze_new.