Add functionality to know and update which avatars are in which squares
The maze structure needs to know which avatars are in which squares mostly
so it can draw them there; however, there are also some benefits to keeping track of
the avatars' positions in the shared maze struct. The function is_collision is
available outside of the module, though get_num_avatars_here is not, which may need
to change if, outside of mazestruct, it's necessary to know not only whether a
collision has occurred but also how many have collided, especially if it is
necessary to iterate through all the avatars on the given collision square.