Clear terminal before printing UI
GTK has proven a little too finnicky, so instead we'll be directly printing to
terminal. To make the maze always print in the same place each time, clearing
the console each time we call draw_maze. This should make the UI appear animated.

I have also put in a one second sleep at the beginning of each drawing, to prevent it going too quickly to see, especially for testing. It is likely that we will want to remove this upon integration with the other modules.