Use update() instead of system("clear") for UI
System(clear) is fine for small mazes, but it gets extremely
flickery in larger ones. Using update() and gotoxy(0,0) gets
rid of the flickeriness, though it does have a few side effects,
including a shimmer of a cursor that doesn't go away when I call
curs_set(0) as well as ascii residue below and sometimes to the
right of the updating maze.